[ads] Add SmartNTT virtual prefs support on iOS
