+++ browser/brave_ads/ads_service_factory.cc
@@ -15,6 +15,7 @@
 #include "brave/browser/brave_ads/device_id/device_id_impl.h"
 #include "brave/browser/brave_ads/services/bat_ads_service_factory_impl.h"
 #include "brave/browser/brave_ads/tooltips/ads_tooltips_delegate_impl.h"
+#include "brave/browser/brave_ads/virtual_pref_provider_delegate.h"
 #include "brave/browser/brave_browser_process.h"
 #include "brave/browser/brave_rewards/rewards_service_factory.h"
 #include "brave/browser/brave_rewards/rewards_util.h"
@@ -98,6 +99,7 @@ AdsServiceFactory::BuildServiceInstanceForBrowserContext(
   return std::make_unique<AdsServiceImpl>(
       std::move(delegate), profile->GetPrefs(),
       g_browser_process->local_state(),
+      std::make_unique<VirtualPrefProviderDelegate>(*profile),
       profile->GetDefaultStoragePartition()
           ->GetURLLoaderFactoryForBrowserProcess(),
       brave::GetChannelName(), profile->GetPath(), CreateAdsTooltipsDelegate(),
